terraform: Dev environment based on terraform
This change introduces a new development environment based on guestfs
and terraform (with terraform-provider-libvirt plugin). So far, it
brings a single-node Kubernetes environment on a libvirt qemu+kvm
virtual machine.

After this change, the full and easiest procedure to build lockc and run
it in a VM consists of the following steps.

Building the project. The first possible method is to build it in a
container:

  ./containerized-build.sh

The second possible method is to do it directly with Meson:

  CC=clang meson build

The container method is recommended - it's easier and ensures that there
is no mismatch of dependency (rust, cargo, clang, llvm, libbpf, bpftool)
versions between your host, CI and other developers' environments.

The next step is building the base VM qcow2 image. We use guestfs for
that. Therefore, ensure that you have `guestfs-tools` installed on your
host. If so, you can proceed with:

  pushd contrib/guestfs
  ./build.sh
  popd

If the `build.sh` script ran successfully, `contrib/guestfs` directory
should contain the `lockc-base.qcow2` image.

The `build.sh` script can be also used to create a VM image with a
custom kernel if there is a need for kernel testing. You can optionally
provide a path to your kernel source tree. Please note that the kernel
should be already build on the host with `make`. Our guestfs scripts do
only `make modules_install install` to install the kernel image and
modules inside a VM. Installing the custom kernel is enabled by using
the `CUSTOM_KERNEL` environment variable. Its value has to be set to
`true`. By default, the script assumes that your kernel tree is in
`~/linux` directory. You can provide a custom path by another
environment variable - `KERNEL_SOURCE`. Examples of usage:

  CUSTOM_KERNEL=true ./build.sh
  CUSTOM_KERNEL=true KERNEL_SOURCE=${HOME}/my-git-repos/linux ./build.sh

If you already used `build.sh` once and you would like to inject a
custom kernel into already build qcow2 image, there is a separate script
- `reinstall-custom-kernel.sh`. It takes an optional `KERNEL_SOURCE`
environment variable. Examples of usage:

  ./reinstall-custom-kernel.sh
  KERNEL_SOURCE=${HOME}/my-git-repos/linux ./reinstall-custom-kernel.sh

VMs which we are going to run are using 9p to mount the source tree. To
ensure that those mounts are going to work correctly, open the
`/etc/libvirt/qemu.conf` file and ensure that the following options
are present there:

  user = "root"
  group = "root"
  dynamic_ownership = 0

If you had to edit the configuration, save the file and restart
libvirt:

  sudo systemctl restart libvirt

After that, we can go to the directory with Terraform environment:

  cd contrib/terraform/libvirt

The important step is to configure the variables. First, copy the
example config file:

  cp terraform.tfvars.json.example terraform.tfvars.json

Then open the `terraform.tfvars.json` config file with your favorite
editor and put your SSH key in the `authorized_keys` field.

After saving the file, we can finally start the VM by doing:

  terraform init
  terraform apply

The successful execution of Terraform should result in the output with
an IP address of the VM. For example:

  Apply complete! Resources: 5 added, 0 changed, 0 destroyed.

  Outputs:

  ip_control_planes = {
    "lockc-control-plane-0" = "10.16.0.116"
  }

Then we can SSH to that VM, using its IP address and `opensuse` user:

  ssh opensuse@10.16.0.116

Inside the VM we can check whether Kubernetes is running:

  # kubectl get pods -A
  NAMESPACE     NAME                                            READY
  STATUS    RESTARTS   AGE
  kube-system   coredns-78fcd69978-lvshz                        0/1
  Running   0          7s
  kube-system   coredns-78fcd69978-q874s                        0/1
  Running   0          7s
  kube-system   etcd-lockc-control-plane-0                      1/1
  Running   0          11s
  kube-system   kube-apiserver-lockc-control-plane-0            1/1
  Running   0          10s
  kube-system   kube-controller-manager-lockc-control-plane-0   1/1
  Running   0          11s
  kube-system   kube-proxy-p7nrd                                1/1
  Running   0          7s
  kube-system   kube-scheduler-lockc-control-plane-0            1/1
  Running   0          11s

And whether lockc is running. The main service can be checked by:

  systemctl status lockcd

We can check also whether lockc's BPF programs are running:

  # bpftool prog list
  35: tracing  name sched_process_f  tag b3c2c2a08effc879  gpl
  	loaded_at 2021-08-10T12:23:55+0000  uid 0
  	xlated 1528B  jited 869B  memlock 4096B  map_ids 3,2
  	btf_id 95
  36: lsm  name clone_audit  tag 33a5e8a5da485fd4  gpl
  	loaded_at 2021-08-10T12:23:55+0000  uid 0
  	xlated 1600B  jited 899B  memlock 4096B  map_ids 3,2
  	btf_id 95
  37: lsm  name syslog_audit  tag 80d655f557922055  gpl
  	loaded_at 2021-08-10T12:23:55+0000  uid 0
  	xlated 1264B  jited 714B  memlock 4096B  map_ids 3,2
  	btf_id 95
  [...]

And whether it registers containers. Directories inside
`/sys/fs/bpf/lockc` represent timestamps of lockcd launch, so it will be
different than in the following example.

  # bpftool map dump pinned /sys/fs/bpf/lockc/1628598193/map_containers
  [{
          "key": 4506,
          "value": {
              "policy_level": "POLICY_LEVEL_PRIVILEGED"
          }
  [...]

Fixes: #23
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>